% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imf_countries.R
\name{imf_countries}
\alias{imf_countries}
\title{Extracts a tibble of IMF - World Economic Outlook by countries}
\usage{
imf_countries()
}
\value{
Returns a tible
}
\description{
Extracts a tibble of IMF - World Economic Outlook by countries
}
