% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imf_country_groups.R
\name{imf_country_groups}
\alias{imf_country_groups}
\title{Extracts a tibble of IMF - World Economic Outlook by country groups}
\usage{
imf_country_groups()
}
\value{
Returns a tible
}
\description{
Extracts a tibble of IMF - World Economic Outlook by country groups
}
