% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMFgroups.R
\docType{data}
\name{IMFgroups}
\alias{IMFgroups}
\title{IMF - World Economic Outlook (WEO) data by groups}
\format{
A tibble with 83200 observations and 11 variables:
\describe{
  \item{weo_country_group_code}{World Economic Outlook (WEO) country group code.}
  \item{weo_subject_code}{World Economic Outlook subject code.}
  \item{country_group_name}{Country group name.}
  \item{subject_descriptor}{Description of the subject.}
  \item{subject_notes}{Subject note.}
  \item{units}{Unit of measurement.}
  \item{scale}{Scale of measurement.}
  \item{country_series_specific_notes}{Country's series specific notes.}
  \item{estimates_start_after}{Year when estimates start after.}
  \item{year}{Year corresponding to values.}
  \item{values}{Values corresponding to years.}

}
}
\source{
\href{https://www.imf.org/}{IMFgroups}
}
\usage{
IMFgroups
}
\description{
Description of variables in the `IMFgroups` tibble
}
\keyword{datasets}
