% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weo_country_code.R
\docType{data}
\name{weo_country_code}
\alias{weo_country_code}
\title{IMF - World Economic Outlook (WEO) metadata by country code}
\format{
A tibble with 196 observations and 3 variables:
\describe{
  \item{weo_country_code}{Country code.}
  \item{iso}{ISO code.}
  \item{country}{Country name.}

}
}
\source{
\href{https://www.imf.org/}{weo_country_code}
}
\usage{
weo_country_code
}
\description{
Description of variables in the `weo_country_code` tibble
}
\keyword{datasets}
