% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weo_country_group_code.R
\docType{data}
\name{weo_country_group_code}
\alias{weo_country_group_code}
\title{IMF - World Economic Outlook (WEO) metadata by country group code}
\format{
A tibble with 13 observations and 2 variables:
\describe{
  \item{weo_country_group_code}{Country group code.}
  \item{country_group_name}{Country group name.}

}
}
\source{
\href{https://www.imf.org/}{weo_country_group_code}
}
\usage{
weo_country_group_code
}
\description{
Description of variables in the `weo_country_group_code` tibble
}
\keyword{datasets}
