% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weo_subject_code.R
\docType{data}
\name{weo_subject_code}
\alias{weo_subject_code}
\title{IMF - World Economic Outlook (WEO) metadata by subject code}
\format{
A tibble with 44 observations and 5 variables:
\describe{
  \item{weo_subject_code}{World Economic Outlook subject code.}
  \item{subject_descriptor}{Description of the subject.}
  \item{subject_notes}{Subject note.}
  \item{units}{Unit of measurement.}
  \item{scale}{Scale of measurement.}

}
}
\source{
\href{https://www.imf.org/}{weo_subject_code}
}
\usage{
weo_subject_code
}
\description{
Description of variables in the `weo_subject_code` tibble
}
\keyword{datasets}
